perf(site/src/pages/AgentsPage/components): isolate live chat tail updates
