+++ site/src/pages/AgentsPage/components/AgentDetail/ConversationTimeline.stories.tsx
@@ -1,22 +1,8 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import type * as TypesGen from "api/typesGenerated";
-import {
-	expect,
-	screen,
-	spyOn,
-	userEvent,
-	waitFor,
-	within,
-} from "storybook/test";
+import { expect, spyOn, userEvent, within } from "storybook/test";
 import { ConversationTimeline } from "./ConversationTimeline";
 import { parseMessagesWithMergedTools } from "./messageParsing";
-import {
-	buildLiveStatus,
-	buildReconnectState,
-	buildRetryState,
-	buildStreamRenderState,
-	textResponseStreamParts,
-} from "./storyFixtures";
 
 // 1×1 solid coral (#FF6B6B) PNG encoded as base64.
 const TEST_PNG_B64 =
@@ -76,27 +62,7 @@ const mockTextAttachmentFetch = () => {
 const defaultArgs: Omit<
 	React.ComponentProps<typeof ConversationTimeline>,
 	"parsedMessages"
-> = {
-	isEmpty: false,
-	streamState: null,
-	streamTools: [],
-	liveStatus: buildLiveStatus(),
-	subagentTitles: new Map(),
-	subagentStatusOverrides: new Map(),
-};
-
-const retryThenResumeMessages = buildMessages([
-	{
-		...baseMessage,
-		id: 1,
-		role: "user",
-		content: [
-			{ type: "text", text: "Please try again if the provider flakes." },
-		],
-	},
-]);
-
-const retryThenResumedStream = buildStreamRenderState(textResponseStreamParts);
+> = {};
 
 const meta: Meta<typeof ConversationTimeline> = {
 	title: "pages/AgentsPage/AgentDetail/ConversationTimeline",
@@ -414,53 +380,6 @@ export const UserMessageWithImagesAndFileRefs: Story = {
 	},
 };
 
-/** Usage-limit errors render as an info alert with analytics access. */
-export const UsageLimitExceeded: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: [],
-		liveStatus: buildLiveStatus({
-			persistedError: {
-				kind: "usage-limit",
-				message:
-					"You've used $50.00 of your $50.00 spend limit. Your limit resets on July 1, 2025.",
-			},
-		}),
-		subagentTitles: new Map(),
-		subagentStatusOverrides: new Map(),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(canvas.getByText(/spend limit/i)).toBeVisible();
-		const link = canvas.getByRole("link", { name: /view usage/i });
-		expect(link).toBeVisible();
-		expect(link).toHaveAttribute("href", "/agents/analytics");
-	},
-};
-
-/** Non-usage errors must not show the usage CTA. */
-export const GenericErrorDoesNotShowUsageAction: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: [],
-		liveStatus: buildLiveStatus({
-			persistedError: {
-				kind: "generic",
-				message: "Provider request failed.",
-			},
-		}),
-		subagentTitles: new Map(),
-		subagentStatusOverrides: new Map(),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(canvas.getByText(/provider request failed/i)).toBeVisible();
-		expect(
-			canvas.queryByRole("link", { name: /view usage/i }),
-		).not.toBeInTheDocument();
-	},
-};
-
 /** File references render inline with text, matching the chat input style. */
 export const UserMessageWithInlineFileRef: Story = {
 	args: {
@@ -611,232 +530,5 @@ export const StickyUserMessageStructure: Story = {
 		const canvas = within(canvasElement);
 		expect(canvas.getByText("First prompt")).toBeVisible();
 		expect(canvas.getByText("Second prompt")).toBeVisible();
-	},
-};
-
-/** Retry errors render the shared mux-style callout with reason text. */
-export const RetryWithReason: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: [],
-		liveStatus: buildLiveStatus({
-			retryState: buildRetryState({ attempt: 2 }),
-			isAwaitingFirstStreamChunk: true,
-		}),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(
-			canvas.getByRole("heading", { name: /retrying request/i }),
-		).toBeVisible();
-		expect(canvas.getByText(/transient upstream failure/i)).toBeVisible();
-		expect(canvas.getByText("generic")).toBeVisible();
-		expect(canvas.getByText(/attempt 2/i)).toBeVisible();
-	},
-};
-
-/** Reconnecting keeps partial streamed output visible without a terminal banner. */
-export const ReconnectingKeepsPartialOutputVisible: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: retryThenResumeMessages,
-		streamState: retryThenResumedStream.streamState,
-		streamTools: retryThenResumedStream.streamTools,
-		liveStatus: buildLiveStatus({
-			streamState: retryThenResumedStream.streamState,
-			reconnectState: buildReconnectState({
-				attempt: 2,
-				delayMs: 2000,
-				retryingAt: "2099-01-01T00:00:00.000Z",
-			}),
-		}),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(canvas.getByText(/storybook streamed answer/i)).toBeVisible();
-		expect(
-			canvas.getByRole("heading", { name: /reconnecting/i }),
-		).toBeVisible();
-		expect(canvas.getByText(/chat stream disconnected/i)).toBeVisible();
-		expect(canvas.queryByText("generic")).not.toBeInTheDocument();
-		expect(
-			canvas.queryByRole("heading", { name: /request failed/i }),
-		).not.toBeInTheDocument();
-	},
-};
-
-/** Overloaded terminal errors expose provider metadata and status links. */
-export const TerminalOverloadedError: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: [],
-		liveStatus: buildLiveStatus({
-			streamError: {
-				kind: "overloaded",
-				message: "Anthropic is currently overloaded. Please try again shortly.",
-				provider: "anthropic",
-				retryable: true,
-				statusCode: 529,
-			},
-		}),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(
-			canvas.getByRole("heading", { name: /service overloaded/i }),
-		).toBeVisible();
-		expect(canvas.getByText("overloaded")).toBeVisible();
-		expect(canvas.getByText(/http 529/i)).toBeVisible();
-		expect(canvas.getByRole("link", { name: /status/i })).toBeVisible();
-	},
-};
-
-/** Store-driven stream errors render the shared terminal callout. */
-export const StreamErrorTerminalFailure: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: retryThenResumeMessages,
-		liveStatus: buildLiveStatus({
-			streamError: {
-				kind: "overloaded",
-				message: "Anthropic is currently overloaded. Please try again shortly.",
-				provider: "anthropic",
-				retryable: true,
-				statusCode: 529,
-			},
-		}),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(
-			canvas.getByText(/please try again if the provider flakes/i),
-		).toBeVisible();
-		const headings = screen.getAllByRole("heading", {
-			name: /service overloaded/i,
-		});
-		expect(headings).toHaveLength(1);
-		expect(headings[0]).toBeVisible();
-		expect(canvas.getByText("overloaded")).toBeVisible();
-		expect(canvas.getByText(/http 529/i)).toBeVisible();
-		const statusLink = canvas.getByRole("link", { name: /status/i });
-		expect(statusLink).toBeVisible();
-		expect(statusLink).toHaveAttribute("href", "https://status.anthropic.com");
-	},
-};
-
-/** Persisted generic errors yield to live streamed content. */
-export const PersistedGenericErrorDoesNotOverrideStreaming: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: retryThenResumeMessages,
-		streamState: retryThenResumedStream.streamState,
-		streamTools: retryThenResumedStream.streamTools,
-		liveStatus: buildLiveStatus({
-			streamState: retryThenResumedStream.streamState,
-			persistedError: {
-				kind: "generic",
-				message: "Stale persisted error.",
-			},
-		}),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		await waitFor(() => {
-			expect(canvas.getByText(/storybook streamed answer/i)).toBeVisible();
-		});
-		expect(
-			canvas.queryByRole("heading", { name: /request failed/i }),
-		).not.toBeInTheDocument();
-	},
-};
-
-/** Failed streams keep partial output visible alongside the terminal callout. */
-export const FailedStreamKeepsPartialOutputVisible: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: retryThenResumeMessages,
-		streamState: retryThenResumedStream.streamState,
-		streamTools: retryThenResumedStream.streamTools,
-		liveStatus: buildLiveStatus({
-			streamState: retryThenResumedStream.streamState,
-			streamError: {
-				kind: "generic",
-				message: "Provider request failed.",
-			},
-		}),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(canvas.getByText(/storybook streamed answer/i)).toBeVisible();
-		expect(
-			canvas.getByRole("heading", { name: /request failed/i }),
-		).toBeVisible();
-		expect(canvas.getByText(/provider request failed/i)).toBeVisible();
-	},
-};
-
-/** Generic terminal errors keep the shared layout without usage/status CTAs. */
-export const TerminalGenericError: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: [],
-		liveStatus: buildLiveStatus({
-			persistedError: {
-				kind: "generic",
-				message: "Provider request failed.",
-			},
-		}),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		expect(
-			canvas.getByRole("heading", { name: /request failed/i }),
-		).toBeVisible();
-		expect(canvas.getByText("generic")).toBeVisible();
-		expect(canvas.getByText(/provider request failed/i)).toBeVisible();
-		expect(
-			canvas.queryByRole("link", { name: /status/i }),
-		).not.toBeInTheDocument();
-		expect(
-			canvas.queryByRole("link", { name: /view usage/i }),
-		).not.toBeInTheDocument();
-	},
-};
-
-/** Delayed first chunks keep the thinking placeholder until the grace period expires. */
-export const DelayedFirstChunk: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: retryThenResumeMessages,
-		liveStatus: buildLiveStatus({ isAwaitingFirstStreamChunk: true }),
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		const matches = canvas.getAllByText(/thinking\.\.\./i);
-		expect(matches.length).toBeGreaterThanOrEqual(1);
-		expect(matches.some((el) => el.closest("[aria-hidden]") === null)).toBe(
-			true,
-		);
-		expect(
-			canvas.queryByRole("heading", { name: /retrying request/i }),
-		).not.toBeInTheDocument();
-	},
-};
-
-/** Once streaming resumes, the retry callout disappears and content remains. */
-export const RetryThenResumedStreaming: Story = {
-	args: {
-		...defaultArgs,
-		parsedMessages: retryThenResumeMessages,
-		...retryThenResumedStream,
-	},
-	play: async ({ canvasElement }) => {
-		const canvas = within(canvasElement);
-		await waitFor(() => {
-			expect(canvas.getByText(/storybook streamed answer/i)).toBeVisible();
-		});
-		expect(
-			canvas.queryByRole("heading", { name: /service overloaded/i }),
-		).not.toBeInTheDocument();
 	},
 };
